+++ R/get_ci.R
@@ -5,7 +5,8 @@ get_ci <- function(
     df = NULL,
     draws = NULL,
     vcov = TRUE,
-    overwrite = FALSE) {
+    overwrite = FALSE,
+    ...) {
 
     if (!is.null(draws)) {
         out <- get_ci_draws(
@@ -38,11 +39,16 @@ get_ci <- function(
     }
 
     if (!"p.value" %in% colnames(x) || isTRUE(overwrite)) {
+        if (!"df" %in% colnames(x) && is.numeric(df)) {
+            x[["df"]] <- df
+        }
+        
         if ("df" %in% colnames(x)) {
             x[["p.value"]] <- 2 * stats::pt(-abs(x$statistic), df = x[["df"]])
         # get_predicted does not save DF and does not compute p.value. We try
         # to extract df in predictions(), but this does not always work
         # (e.g., with hypothesis). When we don't have DF, normal p.value is misleading.
+
         } else if (!identical(vcov, "satterthwaite") || !identical(vcov, "kenward-roger")) {
             x[["p.value"]] <- 2 * stats::pnorm(-abs(x$statistic))
         }