+++ R/sanity_dots.R
@@ -36,7 +36,7 @@ sanity_dots <- function(model, calling_function = NULL, ...) {
     valid[["rlmerMod"]] <- c("re.form", "allow.new.levels")
     valid[["gamlss"]] <- c("what", "safe") #gamlss
 
-    white_list <- c("conf.int", "modeldata", "contrast_factor", "contrast_numeric", "internal_call")
+    white_list <- c("conf.int", "modeldata", "contrast_factor", "contrast_numeric", "internal_call", "df")
 
     model_class <- class(model)[1]